fixing anti raids function

+++ commands/clear.js
@@ -7,8 +7,12 @@ module.exports = {
         msg.channel.send("please write how many messages to delete like !clear amount");
         return;
       }
+      if (args[1]>100){
+        msg.channel.send("i can't delete more then 100 messages")
+        return;
+      }
       msg.channel.bulkDelete(parseInt(args[1])+1)
-      .then(msg.channel.send("seccessfully deleted " + args[1] + " messages")).catch(error => console.error(error));
+      .catch(error => console.error(error)).then(msg.channel.send("seccessfully deleted " + args[1] + " messages"));
     }
   }
 };
